Feature/testing aditional functions

+++ src/main/java/com/api/scholarships/services/implementation/UserServiceImp.java
@@ -83,8 +83,10 @@ public class UserServiceImp implements UserService {
 
   @Override
   public void delete(Long id) {
-    // TODO:verify if an user with a company it's possible delete
     User userFound = getById(id);
+    if(userFound.getCompany()!= null){
+      throw new BadRequestException(Messages.MESSAGE_CANNOT_DELETE_USER);
+    }
     userRepository.delete(userFound);
   }
